[TODO]variant baseline

+++ include/memory_node.h
@@ -177,7 +177,7 @@ private:
     FreeQueueManager *free_queue_manager_;
     uint8_t running;
     uint32_t global_rkey_;
-    std::atomic<int> reg_size_;
+    std::atomic<uint64_t> reg_size_;
     std::atomic<uint64_t> heap_pointer_;
 
     std::mutex m_mutex_;
